+++ Realm/Realm/DataBinding/RealmObjectTypeDelegator.cs
@@ -45,7 +45,18 @@ namespace Realms.DataBinding
                 return null;
             }
 
-            if (_schema.Any(p => p.ManagedName == name))
+            if (_schema != null)
+            {
+                if (_schema.Any(p => p.ManagedName == name))
+                {
+                    return _propertyCache.GetOrAdd(name, n => new WovenPropertyInfo(result));
+                }
+
+                return result;
+            }
+
+            // Schema is null only for woven unmanaged objects
+            if (result?.GetCustomAttribute<WovenPropertyAttribute>() != null)
             {
                 return _propertyCache.GetOrAdd(name, n => new WovenPropertyInfo(result));
             }